docs: sync README and docs with latest architecture and scripting changes
docs(readme): add recent development highlights from latest commits

docs(scripting): align runtime command layout to moongate_data/scripts/commands/gm

docs(spatial): document light-cycle split into ILightService/LightService

docs(persistence): replace MemoryPack references with MessagePack-CSharp source-gen

chore(format): include repository-wide formatting updates

+++ src/Moongate.Network.Packets/Outgoing/Entity/MobileMovingPacket.cs
@@ -37,7 +37,7 @@ public sealed class MobileMovingPacket : BaseGameNetworkPacket
         writer.Write((short)Mobile.Location.Y);
         writer.Write((sbyte)Mobile.Location.Z);
         writer.Write((byte)Mobile.Direction);
-        writer.Write((short)Mobile.SkinHue);
+        writer.Write(Mobile.SkinHue);
         writer.Write(Mobile.GetPacketFlags(StygianAbyss));
         writer.Write((byte)Mobile.Notoriety);
     }
